refactor(Inicio RUP): se modulariza y se llevan métodos a service
