+++ src/app/components/turnos/turnos-rup/rup-fuera-de-agenda.component.ts
@@ -1,12 +1,12 @@
 
 import { Component, OnInit, Input } from '@angular/core';
-import { PrestacionesService } from '../../../services/prestaciones.service';
+import { PrestacionesService } from '../../../modules/rup/services/prestaciones.service';
 
 @Component({
     selector: 'rup-fuera-de-agenda',
     templateUrl: 'rup-fuera-de-agenda.html',
     styleUrls: [
-        '../../core/_rup.scss'
+        './turnos-rup.scss'
     ]
 })
 export class RUPFueraDeAgendaComponent implements OnInit {
