+++ src/app/components/turnos/turnos-rup/rup-fuera-de-agenda.html
@@ -2,7 +2,7 @@
     <!-- {{ prestacion | json }} -->
     <div class="prestacion">
         <div class="rup-icon">
-            <i class="icon icon-rup-semantic-plan"></i>
+            <i class="adi adi-rup-semantic-plan"></i>
         </div>
         <div class="d-flex flex-column info">
             <div class="d-flex flex-column ml-1" *ngIf="prestacion?.paciente">
