+++ src/app/components/turnos/turnos-rup/rup-turnos.html
@@ -31,9 +31,6 @@
                                [ngClass]="{'text-white': agendaSeleccionada == agenda, 'text-primary': agendaSeleccionada !== agenda, 'sm': (turno?.estado == 'disponible' || turno?.estado === 'turnoDoble')}">
                     </plex-icon>
                     <strong>{{ turno.horaInicio | date: 'HH:mm' }}
-                        <span *ngIf="turno && turno.nota">
-                            <i title="{{turno.nota}}" class="text-warning warning mdi mdi-message"></i>
-                        </span>
                     </strong>
                 </div>
             </div>
@@ -41,6 +38,9 @@
                 <div>
                     <ng-container *ngIf="turno.paciente?.id">
                         <b>{{ turno.paciente | paciente }}</b>
+                        <span *ngIf="turno && turno.nota">
+                            <i title="{{turno.nota}}" class="text-warning warning mdi mdi-message"></i>
+                        </span>
                         <small class="d-flex">
                             <span>
                                 <span *ngIf="turno.paciente.documento !== ''">
@@ -114,23 +114,22 @@
             </div>
             <div justify="end" *ngIf="turno.tipoPrestacion && turno.paciente?.id">
 
-
-
                 <!--Iniciar una prestacion a partir de una solicitud pendiente (es decir una prestacion en estado pendiente) -->
-                <!-- <plex-button class="mr-1" size="sm"
+                <plex-button class="mr-1" size="sm"
                              *ngIf="turno.paciente && turno.estado !== 'suspendido'  && turno.prestacion && turno.prestacion.estados[turno.prestacion.estados.length - 1].tipo === 'pendiente' && servicioPrestacion.tienePermisos(turno.tipoPrestacion, turno.prestacion) && servicioPrestacion.verificarAsistencia(turno)"
                              type="success"
-                             (click)="ejecutarPrestacionPendiente(turno.prestacion.id,turno.paciente, turno.tipoPrestacion, turno)"
+                             (click)="servicioPrestacion.ejecutarPrestacionPendiente(turno.prestacion.id,turno.paciente, turno.tipoPrestacion, turno)"
                              [disabled]="esFutura" title="Iniciar Prestación">
                     <i class="mdi mdi-plus-circle mdi-24px"></i>
-                </plex-button> -->
+                </plex-button>
 
                 <!--Una agenda futura muestra en boton deshabilitado -->
-                <!-- <plex-button size="sm"
-                                 *ngIf="esFutura && turno.paciente && turno.estado !== 'suspendido'  && turno.prestacion && turno.prestacion.estados[turno.prestacion.estados.length - 1].tipo === 'pendiente' && servicioPrestacion.tienePermisos(turno.tipoPrestacion, turno.prestacion) && servicioPrestacion.verificarAsistencia(turno)"
-                                 icon="plus-circle" type="disabled"
-                                 title="Disponible {{diaAgenda(agendaSeleccionada)}}" titlePosition="top">
-                    </plex-button> -->
+                <plex-button size="sm"
+                             *ngIf="esFutura && turno.paciente && turno.estado !== 'suspendido'  && turno.prestacion && turno.prestacion.estados[turno.prestacion.estados.length - 1].tipo === 'pendiente' && servicioPrestacion.tienePermisos(turno.tipoPrestacion, turno.prestacion) && servicioPrestacion.verificarAsistencia(turno)"
+                             icon="plus-circle" type="disabled"
+                             title="Disponible {{ servicioPrestacion.diaAgenda(agendaSeleccionada) }}"
+                             titlePosition="top">
+                </plex-button>
 
                 <!-- Iniciar Prestación -->
                 <plex-button class="mr-1" size="sm"
@@ -141,11 +140,12 @@
                     <i class="mdi mdi-plus-circle mdi-24px"></i>
                 </plex-button>
 
-                <!-- <plex-button size="sm"
+                <plex-button size="sm"
                              *ngIf="esFutura && agendaSeleccionada.estado !== 'auditada' && turno.estado !== 'suspendido' && turno.paciente && !turno.prestacion && servicioPrestacion.verificarAsistencia(turno)"
                              icon="plus-circle" type="disabled"
-                             title="Disponible {{diaAgenda(agendaSeleccionada)}}" titlePosition="top">
-                </plex-button> -->
+                             title="Disponible {{ servicioPrestacion.diaAgenda(agendaSeleccionada) }}"
+                             titlePosition="top">
+                </plex-button>
 
                 <plex-button class="mr-1" size="sm"
                              *ngIf="turno.paciente && turno.estado !== 'suspendido' && turno.prestacion && turno.prestacion.estados[turno.prestacion.estados.length - 1].tipo === 'ejecucion' && servicioPrestacion.tienePermisos(turno.tipoPrestacion, turno.prestacion) && servicioPrestacion.verificarAsistencia(turno)"
@@ -179,17 +179,18 @@
             </div>
             <div justify="end" *ngIf="agendaSeleccionada.tipoPrestaciones[0].noNominalizada">
                 <!--TODO -> revisar cuando se defina el estado final de las agendas no nominalizadas agendaSeleccionada.estado !== 'auditada'-->
-                <!-- <plex-button size="sm"
+                <plex-button size="sm"
                              *ngIf="!turno.asistencia &&  !esFutura && agendaSeleccionada.estado !== 'auditada' && turno.estado !== 'suspendido' && !turno.prestacion && servicioPrestacion.tienePermisos(turno.tipoPrestacion, turno.prestacion)"
                              icon="plus-circle" type="success"
-                             (click)="iniciarPrestacionNoNominalizada(turno.tipoPrestacion, turno)">
+                             (click)="servicioPrestacion.iniciarPrestacionNoNominalizada(turno.tipoPrestacion, turno)">
                 </plex-button>
 
                 <plex-button size="sm"
                              *ngIf="!turno.asistencia &&  esFutura && agendaSeleccionada.estado !== 'auditada' && turno.estado !== 'suspendido' && !turno.prestacion"
                              icon="plus-circle" type="disabled"
-                             title="Disponible {{diaAgenda(agendaSeleccionada)}}" titlePosition="top">
-                </plex-button> -->
+                             title="Disponible {{ servicioPrestacion.diaAgenda(agendaSeleccionada) }}"
+                             titlePosition="top">
+                </plex-button>
 
                 <plex-button size="sm"
                              *ngIf="turno.estado !== 'suspendido' && turno.prestacion && turno.prestacion.estados[turno.prestacion.estados.length - 1].tipo === 'ejecucion' && servicioPrestacion.tienePermisos(turno.tipoPrestacion,turno.prestacion)"
