+++ src/app/modules/rup/services/prestaciones.service.ts
@@ -11,6 +11,7 @@ import { IPrestacionGetParams } from '../interfaces/prestacionGetParams.interfac
 import { SnomedService } from '../../../services/term/snomed.service';
 import { ReglaService } from '../../../services/top/reglas.service';
 import { Router } from '@angular/router';
+import { IAgenda } from '../../../interfaces/turnos/IAgenda';
 
 
 
@@ -996,6 +997,88 @@ export class PrestacionesService {
                 });
             } else {
                 return false;
+            }
+        });
+    }
+
+    /**
+       * Ejecutar una prestacion que esta en estado pendiente
+    */
+    ejecutarPrestacionPendiente(idPrestacion, paciente, snomedConcept, turno) {
+        let params: any = {
+            op: 'estadoPush',
+            ejecucion: {
+                fecha: turno.horaInicio,
+                registros: [],
+                // organizacion desde la que se solicita la prestacion
+                organizacion: { id: this.auth.organizacion.id, nombre: this.auth.organizacion.nombre }
+            },
+            estado: { tipo: 'ejecucion' }
+        };
+
+        this.plex.confirm('Paciente: <b>' + paciente.apellido + ', ' + paciente.nombre + '.</b><br>Prestación: <b>' + snomedConcept.term + '</b>', '¿Iniciar Prestación?').then(confirmacion => {
+            if (confirmacion) {
+                this.patch(idPrestacion, params).subscribe(prestacion => {
+                    this.router.navigate(['/rup/ejecucion', idPrestacion]);
+                }, (err) => {
+                    this.plex.info('warning', 'No fue posible iniciar la prestación: ' + err, 'ERROR');
+                });
+            } else {
+                return false;
+            }
+        });
+    }
+    // Detecta si "hoy" es el día de la Agenda
+    diaAgenda(agenda: IAgenda) {
+        return moment(agenda.horaInicio).fromNow();
+    }
+
+    iniciarPrestacionNoNominalizada(snomedConcept, turno, agenda) {
+        this.plex.confirm('</b><br>Prestación: <b>' + snomedConcept.term + '</b>', '¿Crear Prestación?').then(confirmacion => {
+            if (confirmacion) {
+                this.crearPrestacion(null, snomedConcept, 'ejecucion', turno.horaInicio, turno).subscribe(prestacion => {
+                    this.routeTo('ejecucion', prestacion.id);
+                    if (agenda && agenda !== 'fueraAgenda') {
+                        let agendaStore = agenda ? agenda : null;
+                        localStorage.setItem('idAgenda', agendaStore.id);
+                    }
+                }, (err) => {
+                    this.plex.info('warning', 'No fue posible crear la prestación', 'ERROR');
+                });
+            } else {
+                return false;
+            }
+        });
+    }
+
+    /**
+     * Recorremos los bloques y los turnos de una agenda
+     * y verifica si hay algun paciente agregado
+     */
+    getCantidadPacientes(agenda): number {
+        let total = 0;
+
+        let lengthBloques = agenda.bloques.length;
+        for (let indexBloque = 0; indexBloque < lengthBloques; indexBloque++) {
+
+            let _turnos = agenda.bloques[indexBloque].turnos.filter(t => {
+                total += (t.paciente && t.paciente.id) ? 1 : 0;
+            });
+        }
+        if (agenda.sobreturnos && agenda.sobreturnos.length > 0) {
+            total = total + agenda.sobreturnos.length;
+        }
+
+
+        return total;
+    }
+
+    getCantidadPacientesAgendas(agendas): number {
+        return agendas.forEach(agenda => {
+            if (this.getCantidadPacientes(agenda) > 0) {
+                return this.getCantidadPacientes(agenda);
+            } else {
+                return 0;
             }
         });
     }
